refactor: changing priority to enum

+++ pumpkin-crates/core/src/engine/state.rs
@@ -360,7 +360,7 @@ impl State {
         let propagator = constructor.create(constructor_context);
 
         pumpkin_assert_simple!(
-            propagator.priority() <= 3,
+            propagator.priority() as u8 <= 3,
             "The propagator priority exceeds 3.
              Currently we only support values up to 3,
              but this can easily be changed if there is a good reason."
